redirect

+++ app/Http/Controllers/TasksController.php
@@ -52,17 +52,16 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        
         $this->validate($request, [
             'status' => 'required|max:10',
             'content' => 'required|max:191',
         ]);
-        
-       $request->user()->tasks()->create([
+            
+        $request->user()->tasks()->create([
             'content' => $request->content,
             'status' => $request->status
         ]);
-        
+            
         return redirect('/');
     }
 
@@ -80,6 +79,7 @@ class TasksController extends Controller
             return view('tasks.show', [
                 'task' => $task,
             ]);
+            
         } else {
             return redirect('/');
         }
@@ -95,10 +95,14 @@ class TasksController extends Controller
     {   
         $task = Task::find($id);
        
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
-       
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+            
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -133,9 +137,13 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('/');
             
-        return redirect('/');
-       
+        } else {
+            return redirect('/');
+        }
     }
 }
